% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapple.R
\name{mapple}
\alias{mapple}
\title{Return value which matched a key in a column}
\usage{
mapple(df, col, key, val)
}
\arguments{
\item{df}{data frame to be searched}

\item{col}{A character. A column from which you search the key.}

\item{key}{A character. A value which matches the condition of col.}

\item{val}{A character. A column name to be returned from data frame.}
}
\value{
A vector.
}
\description{
Return value which matched a key in a column
}
\examples{

df <- tibble::rownames_to_column(mtcars,"kind")
mapple(df,"kind","Mazda RX4","hp")
}
